ENH: move file path creator back to save_powder_output
